added new PNGs, feature and steps file for Name Last column; updated common and look-people-basic steps files
